CSV transport layer
Introduce decimal-degrees (DD) parsing and centralize CSV I/O. Add CoordinateFormat CLI flag and separate InputFormat (CSV). Make Coordinate.dir optional and update DMS/DDM parsing to use Option<char>. Implement dd module with dd_to_dd and DdError, add InvalidDd and InvalidRow errors. Refactor models to RawCoordinate/InputRow, move CSV header validation and row loading into transport/csv::load_csv_rows (with line numbers and strict handling), and update main, parser, and process to use the new plumbing and error types.

+++ src/pipeline/process.rs
@@ -68,7 +68,7 @@ pub fn process_geo(
     geo: &NormalizedGeo,
     id: &mut u64,
     strict: bool,
-    invalid: &mut u64,
+    invalid: &mut usize,
 ) -> Result<(), AppError> {
     // Compute distance.
     let dist_km = round(haversine(geo.a.lat.dd, geo.a.lon.dd, geo.b.lat.dd, geo.b.lon.dd)?, 2);
